pyproject.toml
[project]
name = "ansible-web"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Mike Pettypiece", email = "mike@pettypiece.ca" },
]
dependencies = [
    "ansible>=13.0,<14",
    "ansible-lint>=26.0,<27",
    "flake8>=7.0,<8",
    "molecule>=25.0,<26",
    "molecule-plugins[docker]>=25.0,<26",
    "pytest>=9.0,<10",
    "pytest-testinfra>=10.0,<11",
    "yamllint>=1.38,<2",
]

[tool.uv]
package = false
